Split traffic_mode.hpp into several files.
