+++ openlr/openlr_match_quality/openlr_assessment_tool/points_controller_delegate_base.hpp
@@ -1,0 +1,33 @@
+#pragma once
+
+#include "indexer/feature.hpp"
+
+#include "geometry/point2d.hpp"
+
+#include <vector>
+
+using FeaturePoint = std::pair<FeatureID, size_t>;
+
+/// This class is responsible for collecting junction points and
+/// checking user's clicks.
+class PointsControllerDelegateBase
+{
+public:
+  enum class ClickType
+  {
+    Miss,
+    Add,
+    Remove
+  };
+
+  virtual std::vector<m2::PointD> GetAllJunctionPointsInViewport() const = 0;
+  /// Returns all junction points at a given location in the form of feature id and
+  /// point index in the feature.
+  virtual std::pair<std::vector<FeaturePoint>, m2::PointD> GetCandidatePoints(
+      m2::PointD const & p) const = 0;
+  virtual std::vector<m2::PointD> GetReachablePoints(m2::PointD const & p) const = 0;
+
+  virtual ClickType CheckClick(m2::PointD const & clickPoint,
+                               m2::PointD const & lastClickedPoint,
+                               std::vector<m2::PointD> const & reachablePoints) const = 0;
+};
